% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_geneSet_collection_DESeq2}
\alias{build_geneSet_collection_DESeq2}
\title{Create statistics and gene sets for isomiR and miRNA from XICRA data analyzed with DESeq2}
\usage{
build_geneSet_collection_DESeq2(data_df_given, name, path2store_results)
}
\arguments{
\item{data_df_given}{Dataframe from XICRA::prepare_data containing type, UID, parent column}

\item{name}{tag for store results}

\item{path2store_results}{absolute path to store results produce}
}
\description{
This function allows to parse given dataframe produced by DESeq2 from XICRA results
and generates statistics according to type of isomiR and genes sets in gmt format for
overrepresentation analysis
}
\keyword{XICRA}
