% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_geneSet_stats_simulations}
\alias{create_geneSet_stats_simulations}
\title{Create statistics and gene sets for isomiR and miRNA from XICRA simulations}
\usage{
create_geneSet_stats_simulations(data_df)
}
\arguments{
\item{data_df_given}{Dataframe from XICRA::prepare_data containing type, UID, parent column}
}
\description{
This function allows to parse given dataframe produced XICRA simulations results
and generates statistics according to type of isomiR
}
\keyword{XICRA}
