% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{discard_0_counts}
\alias{discard_0_counts}
\title{Summarize counts by miRNA & variant}
\usage{
discard_0_counts(countsF)
}
\arguments{
\item{dataframe}{obtain from XICRA.stats::prepare_data or XICRA.stats::prepare_counts_by_variant or XICRA.stats::prepare_counts_by_miRNA}
}
\description{
This functions prepares countsfile for DESeq discarding rows with
many 0 for all samples.
}
\keyword{XICRA}
