% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Biotypes.R
\name{get_data}
\alias{get_data}
\title{Get data and summarize for main biotypes}
\usage{
get_data(data_biotypes)
}
\arguments{
\item{data_biotypes}{Results file from XICRA biotype and/or featurecount + modifications}
}
\description{
This function allows to parse given file, from XICRA biotype to produce a dataframe containing
summary statistics for main RNA biotyoes (miRNA, tRNA, piRNA, miscRNA, lincRNA, etc)
}
\keyword{XICRA}
