% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_XICRA}
\alias{parse_XICRA}
\title{Read isomiR data and return: isormiR, miRNA and variant type}
\usage{
parse_XICRA(XICRA_csv_file)
}
\arguments{
\item{file_given}{miRNAexpression table generated by XICRA}
}
\description{
This function allows to read given file produced by XICRA pipeline
containing unique ID and returns isomiR, variant type and miRNA data in
separate dataframes
}
\keyword{XICRA}
