% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Biotypes.R
\name{plot_biotypes}
\alias{plot_biotypes}
\title{Plot RNA biotype data into stack bar plots}
\usage{
plot_biotypes(data_perc, palette_color_name = "Set3", out_folder = NULL)
}
\arguments{
\item{data_perc}{Data converted to 100\% results.}

\item{palette_color_name}{Name of the color palette.}

\item{out_folder}{Output folder to store results.}
}
\description{
This function allows to plot given data from XICRA biotype and parse by get_data()
}
\keyword{XICRA}
