% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_counts_by_miRNA}
\alias{prepare_counts_by_miRNA}
\title{Summarize counts by miRNA}
\usage{
prepare_counts_by_miRNA(dataGiven)
}
\arguments{
\item{dataframe}{obtain from XICRA.stats::prepare_data}
}
\description{
This function allows to summarize isomiR data by miRNA
}
\keyword{XICRA}
