% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_counts_by_variant}
\alias{prepare_counts_by_variant}
\title{Summarize counts by miRNA & variant}
\usage{
prepare_counts_by_variant(dataGiven)
}
\arguments{
\item{dataframe}{obtain from XICRA.stats::prepare_data}
}
\description{
This function allows to summarize isomiR data by miRNA and variant type
}
\keyword{XICRA}
