% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_data}
\alias{prepare_data}
\title{Read isomiR data}
\usage{
prepare_data(file_given)
}
\arguments{
\item{file_given}{miRNAexpression table generated by XICRA}
}
\description{
This function allows to read given file produced by XICRA pipeline
containing unique ID for isomiR, variant type, UID and miRNA name
}
\keyword{XICRA}
